fix(review-eval): finder compare refuses inconsistent leak shapes and empty pairings; renderer split out
A result whose leak.hard is non-empty while suspected is false is refused, as the canonical validator refuses it; two runs with no shared completed pipeline cell are refused instead of rendering zero totals. The text renderer moved to review-eval-finder-render.mjs (73 lines, capped at 200) so the compare module lands at 541 against its 600-line gate.

Refs #2379
